Add run into asteroid inflicts damage

+++ Glitter/Headers/CodeMonkeys/TheGauntlet/Projectile.h
@@ -10,7 +10,7 @@ namespace CodeMonkeys::TheGauntlet
     class Projectile : public Particle, public IInflicter
     {
     public:
-        Projectile(Model3D* model, string name, float total_lifespan, ParticleEmitter* emitter, float inflict_amount);
+        Projectile(Model3D* model, string name, float total_lifespan, ParticleEmitter* emitter, unsigned int inflict_amount);
         virtual Particle* clone();
     };
 }
